requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.7.0
gradio>=4.0.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
ffmpeg-python>=0.2.0
moviepy>=1.0.3
tqdm>=4.65.0 